NotificationsMongoDBv2/requirements.txt
aiosmtplib==2.0.1
anyio==3.7.0
blinker==1.6.2
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
exceptiongroup==1.1.1
fastapi==0.95.2
fastapi-mail==1.2.8
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
pydantic==1.10.8
pymongo==4.3.3
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.3
uvicorn==0.22.0
